Testing aec3 transparency
BUG=

+++ webrtc/modules/audio_processing/aec3/mock/mock_render_delay_controller.h
@@ -27,8 +27,8 @@ class MockRenderDelayController : public RenderDelayController {
   MOCK_METHOD0(Reset, void());
   MOCK_METHOD1(SetDelay, void(size_t render_delay));
   MOCK_METHOD2(GetDelay,
-               size_t(const DownsampledRenderBuffer& render_buffer,
-                      rtc::ArrayView<const float> capture));
+               int(const DownsampledRenderBuffer& render_buffer,
+                   rtc::ArrayView<const float> capture));
   MOCK_CONST_METHOD0(AlignmentHeadroomSamples, rtc::Optional<size_t>());
 };
